requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
SQLAlchemy==2.0.23
geopy==2.4.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.1 